Dissect ack (write)
The ack message is a one-byte message without preamble and end fields.
It will be send after each message with handle 0x0016 (write command),
send to the smart watch.

Anyway, wireshark wants to be flexible and does not differate between
sender and receiver. Thus, every one-byte message is interpreted as
ack write.