Reorder frame dissection: ack -> msg
Previously, frame dissection depends on frame length: If the frame
length was 1, it was dissected as ack frame. For all other lengths, it
was dissected as msg frame.

But that condition is not correct: Frames with length 1 will be tried to
be dissected as ack frame. If the value is not correct, it will now be
dissected as msg frame. The reason is that the last fragmented msg frame
can have a length of 1, ending with the frame end value (0x8f).